requirements.txt
selenium>=4.18.0
gspread>=6.0.0
google-auth>=2.29.0
streamlit>=1.32.0
python-dotenv>=1.0.1
requests>=2.31.0
